Fixed constructors and methods in Order, CustomerRepository, OrderRepository and OrderRespsitoryOffline. Methods had different variable type to that which was being called.

+++ src/main/java/com/qa/tangerinedream/repositories/CustomerRepository.java
@@ -32,7 +32,7 @@ public interface CustomerRepository {
 	
 	public void updatePassword(Long id_, String password_);
 	
-	public void updateCredit(Long id_, float f);
+	public void updateCredit(Long id_, int f);
 	
 	public void updateCreditLimit(Long id_, int creditLimit_);
